Split config module into arguments, secrets, and config file.
