+++ deepwell/src/main.rs
@@ -21,7 +21,7 @@
 #![forbid(unsafe_code)]
 #![deny(missing_debug_implementations)]
 
-//! A web server to expose Wikijump operations via an internal trusted REST API.
+//! A web server to expose Wikijump operations via an internal REST API.
 
 #[macro_use]
 extern crate futures;
@@ -51,13 +51,13 @@ mod services;
 mod utils;
 mod web;
 
-use self::config::Config;
+use self::config::SetupConfig;
 use anyhow::Result;
 
 #[async_std::main]
 async fn main() -> Result<()> {
     // Load the configuration so we can set up
-    let config = Config::load();
+    let SetupConfig { secrets, config } = SetupConfig::load();
 
     // Copy fields we need
     let socket_address = config.address;
@@ -75,11 +75,11 @@ async fn main() -> Result<()> {
 
     // Run migrations, if enabled
     if run_migrations {
-        database::migrate(&config.database_url).await?;
+        database::migrate(&secrets.database_url).await?;
     }
 
     // Set up server state
-    let app_state = api::build_server_state(config).await?;
+    let app_state = api::build_server_state(config, secrets).await?;
 
     // Run seeder, if enabled
     if run_seeder {
